requirements.txt
torch==1.9.0
torchvision==0.10.0
imageio==2.9.0
opencv_python==4.1.2.30
numpy==1.20.3
Pillow==8.4.0
